+++ apps/tamma-elsa/tests/Tamma.Api.Tests/Epic28/LruResolverLeaseAndDiagnosticsTests.cs
@@ -80,7 +80,7 @@ public class LruResolverLeaseAndDiagnosticsTests
         entry.Property("Status").CurrentValue = "active";
         entry.Property("EncryptedConnectionString").CurrentValue =
             Encoding.UTF8.GetBytes(cs);
-        entry.Property("KekVersion").CurrentValue = 1;
+        entry.Property("KekVersion").CurrentValue = (short)1;
         await ctx.SaveChangesAsync();
         return tenantId;
     }
