+++ apps/tamma-elsa/tests/Tamma.Api.Tests/Secrets/KekRotationRetryTests.cs
@@ -229,7 +229,7 @@ public class KekRotationRetryTests
         var entry = ctx.Tenants.Add(tenant);
         entry.Property("Status").CurrentValue = "active";
         entry.Property("EncryptedConnectionString").CurrentValue = envelope;
-        entry.Property("KekVersion").CurrentValue = 1;
+        entry.Property("KekVersion").CurrentValue = (short)1;
         await ctx.SaveChangesAsync();
     }
